refactor: move `[[...type-kind]].page` components into sub-directory
