+++ apps/hash-frontend/src/pages/types/[[...type-kind]].page/types-page-tabs.tsx
@@ -1,8 +1,8 @@
 import { FunctionComponent } from "react";
 
-import { TabLink } from "../../shared/ui/tab-link";
-import { Tabs } from "../../shared/ui/tabs";
-import { TabId } from "./[[...type-kind]].page";
+import { TabLink } from "../../../shared/ui/tab-link";
+import { Tabs } from "../../../shared/ui/tabs";
+import { TabId } from "../[[...type-kind]].page";
 
 const tabIds = [
   "all",
